+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Added
 * Visualization library, including plot_sma.
 * TO DO list.
 * Tiingo stock scraper.
+* Validity check against market day.
 
 Changed
 -------